+++ source/common/src/vector2.cpp
@@ -3,11 +3,11 @@
 #include <vector2.hpp>
 
 // Scalar Operations
-Vector2 Vector2::operator+(const double scalar) const {
+Vector2 Vector2::operator+(const float scalar) const {
   return Vector2(x + scalar, y + scalar);
 }
 
-Vector2 Vector2::operator*(const double scalar) const {
+Vector2 Vector2::operator*(const float scalar) const {
   return Vector2(x * scalar, y * scalar);
 }
 
@@ -20,7 +20,7 @@ Vector2 Vector2::operator*(const Vector2 &other) const {
   return Vector2(x * other.x, y * other.y);
 }
 
-double Vector2::dot(const Vector2 &other) { return x * other.x + other.y * y; }
+float Vector2::dot(const Vector2 &other) { return x * other.x + other.y * y; }
 
 Vector2 &Vector2::operator+=(const Vector2 &other) {
   x += other.x;
@@ -28,15 +28,15 @@ Vector2 &Vector2::operator+=(const Vector2 &other) {
   return *this;
 }
 
-Vector2 &Vector2::operator*=(double scalar) {
+Vector2 &Vector2::operator*=(float scalar) {
   x *= scalar;
   y *= scalar;
   return *this;
 }
 
-double Vector2::magnitude() const { return std::sqrt(x * x + y * y); }
+float Vector2::magnitude() const { return std::sqrt(x * x + y * y); }
 
 Vector2 Vector2::normal() const {
-  double length = sqrtf(x * x + y * y);
+  float length = sqrtf(x * x + y * y);
   return Vector2(x / length, y / length);
 }
